+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Errors from backend now show more detail in the cause of the error
 - Compatibility with Node-RED 3.0.0
+- Example flows from Thinger.io official documentation
 
 ### Fix
 - Device write handling plain text as well as JSON